requirements.txt
streamlit
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.19.0
requests>=2.25.0